Conflict resolved
1. Get current Location from IP
2. Set Temperature as Integer instead of Decimal
3. arg checking update to ignore empty value for Provider or Location
4. --p and -p both support is enabled. Because, dotnet run -p gives error as Project name on mac

+++ nahid/xWeather/Drivers/VisualCrossing/VisualCrossingDriver.cs
@@ -35,7 +35,7 @@ namespace xWeather.Drivers.VisualCrossing
                     {
                         Location = location,
                         Condition = weatherResponse.Current.Condition,
-                        Temperature = (weatherResponse.Current.Temperature - 32) * (5.0 / 9),
+                        Temperature = (int) ((weatherResponse.Current.Temperature - 32) * (5.0 / 9)),
                         Humidity = weatherResponse.Current.Humidity,
                         WindSpeed = weatherResponse.Current.WindSpeed,
                         Pressure = weatherResponse.Current.Pressure,
