+++ nahid/xWeather/Drivers/WeatherData.cs
@@ -10,7 +10,7 @@ namespace xWeather.Drivers
     {
         public string Location { get; set; } = "";
         public string Condition { get; set; } = "";
-        public double Temperature { get; set; }
+        public int Temperature { get; set; }
         public double Humidity { get; set; }
         public double WindSpeed { get; set; }
         public double Pressure { get; set; }
